feat: initial Micronaut implementation (#13)
This commit adds the start of the Micronaut (and Postgres, eventually)
implementation of cautious-engine.

The initial phase exposes two endpoints.

1. GET /addresses
*****************
Returns an array containing 1-100 randomly generated addresses.

Sample Response:
{
  "addresses": [
    {
      "street": "89133 Ullrich Grove",
      "city": "East Stellachester",
      "state": "SD",
      "zip_code": "577##",
      "created_at": "2021-02-17T01:39:45.209795Z"
    },
    {
      "street": "778 Robbie Street",
      "city": "Runolfsdottirstad",
      "state": "MS",
      "zip_code": "387##",
      "created_at": "2021-02-17T01:39:45.211489Z"
    }
  ],
  "count": 2
}

2. GET /addresses/:id
*********************

Returns a single randomly generated address. The `:id` parameter can be any
combination of valid characters, as it is ignored and only used for routing.

Sample Response:
{
  "street": "664 Wilfredo Drives",
  "city": "West Dewaynetown",
  "state": "NY",
  "zip_code": "122##",
  "created_at": "2021-02-17T01:28:58.917311Z"
}